Criada uma classe responsável por abrir e fechar a conexão com o servidor. A conexão vai ser aberta quando o sistema iniciar e fechada quando ele for fechado.
