+++ VoceAluga/src/model/Main.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 
 import javafx.application.Application;
 import javafx.stage.Stage;
+import model.server.MySQLConnector;
 import controller.ScenesManager;
 
 public class Main extends Application {
@@ -13,9 +14,17 @@ public class Main extends Application {
 	
 	@Override
     public void start(Stage primaryStage) throws IOException{
+		MySQLConnector.abrirConexao();
+		
     	manager.mostrarTelaDeLogin();
 //		manager.mostrarResultadosBuscaCliente();
         ScenesManager.primaryStage.show();
+    }
+	
+	@Override
+    public void stop() {
+        // executed when the application shuts down
+        MySQLConnector.fecharConexao();
     }
